requirements.txt
streamlit>=1.24.0
langchain>=0.1.0
langchain-community>=0.0.1
langgraph>=0.0.1
typing-extensions>=4.7.1
python-dotenv>=0.19.0
google-serper>=0.1.0
